requirements.txt
django==1.11.4
django-crispy-forms==1.6.1
Pillow==4.2.1
django-material==0.12.0
django-imagekit==3.3